requirements.txt
instagrapi
holidays
Pillow
beautifulsoup4